Refactored types files to make file naming consistent
