+++ quolance-ui/src/util/hooks/userHooks.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { useQuery } from '@tanstack/react-query';
 import { DATA_Users } from '@/constants/data';
-import { UserType } from '@/constants/types/userTypes';
+import { UserType } from '@/constants/types/user-types';
 
 /*--- Hooks ---*/
 // Switch between number '2' and number '3' to see the different user roles
